Fix Key Up
KEY_UP didn't update cursor correctly:
if head == '\n' key up didn't get onto the new line above. Now when pressing KEY_UP, we check if the node is head and if character is of type '\n' in this case we return nullptr.
KEY_UP modified the pointer even though line was 0:
if head->y was equal to 0 and KEY_UP was pressed, the pointer would still have been modified. Thus, placing the cursor at position in front of head, this was corrected.